feat: changing the way i get the last registers on database
