+++ src/App/Spents/Domain/GetLastSpentsDataRegistersRepository.ts
@@ -1,0 +1,5 @@
+import { SpentsData } from "./SpentsData";
+
+export interface GetLastSpentsDataRegistersRepository {
+    getQueryByLastAddRegisters(data: SpentsData[]):Promise<SpentsData[]>   
+}
