+++ src/App/Spents/Domain/SpentsDataRepository.ts
@@ -1,6 +1,5 @@
 import { SpentsData } from './SpentsData'
 
 export interface SpentsDataRepository {
-    save(data: SpentsData[]): Promise<SpentsData[]>
-    getQueryByLastAddRegisters(data: SpentsData[]):Promise<any>
+    save(data: SpentsData[]): Promise<void>
 }
